requirements.txt
streamlit 
langchain-ollama
langchain-core
langchain
langchain-experimental
langchain-google-genai
google-generativeai
pyvis 
python-dotenv
python-docx
pypdf 
openai
pandas
neo4j
nest-asyncio
Pillow>=10.0.0

# lightgraph